added multisupport for bot-refresh.js
