+++ package.json
@@ -18,6 +18,8 @@
 
         "request": "^2.88.2",
 
+        "stockfish": "^15.0.0",
+
         "winston": "^3.8.1",
 
         "xhr2": "^0.2.1"
